% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_brdados.R
\name{tema_brdados}
\alias{tema_brdados}
\title{Our ggplot2 theme}
\usage{
tema_brdados(tam_fonte = 16, fonte = "sans")
}
\arguments{
\item{tam_fonte}{Set font size used in the plot.}

\item{fonte}{Set main font used in the plot.}
}
\description{
\code{tema_br} applies our ggplot2 theme to a ggplot2 plot object.
}
